Clean up EditFeature
